backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
flask-smorest==0.45.0
marshmallow==3.23.2
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.1
werkzeug==3.1.3
pytest==8.3.4
boto3==1.39.11
Flask-Limiter==3.12
PyJWT==2.10.1
